requirements.txt
asgiref==3.8.1
blinker==1.8.2
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
DateTime==5.5
distlib==0.3.9
Django==5.1.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
google-auth==2.35.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.2
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.2
razorpay==1.4.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
SQLAlchemy==2.0.35
sqlalchemy-orm==1.2.10
sqlparse==0.5.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.0.4
zope.interface==7.0.3
